Move ORCA data explorer to packages/explorer/data-explorer
- Move docs/schema-reference.html → packages/explorer/data-explorer/data-explorer.html
- Move packages/contracts/scripts/export-design-reference.js → packages/explorer/data-explorer/build.js
- Update build.js: read domain list from config.yaml, colors from project palette,
  __dirname-relative defaults so no CLI args needed, fix pre-existing crash on
  non-string prop.type values
- Add build:data-explorer to explorer package; wire into unified build script
  (npm run build rebuilds context-map, service-blueprints, and data-explorer)
- Remove design:reference from root package.json
- Remove stale design reference staleness check from preflight.sh
